requirements.txt
numpy
torch
pandas
matplotlib
seaborn
argparse
Pillow
scikit-learn
transformers
torchvision